fixed build
--HG--
extra : convert_revision : svn%3A39bc706e-5318-0410-9160-8a85361fbb7c/trunk%401764

+++ extensions/cstrike/extension.cpp
@@ -206,7 +206,7 @@ bool CStrike::ProcessCommandTarget(cmd_target_info_t *info)
 		{
 			continue;
 		}
-		if (pInfo->GetTeamIndex() != team_index)
+		if (pInfo->GetTeamIndex() != (int)team_index)
 		{
 			continue;
 		}
